+++ src/main/java/com/coderhouse/interfaces/ProductRestInterface.java
@@ -8,13 +8,13 @@ public interface ProductRestInterface {
 	
 	public List<Product> getAllProducts();
 	
-	public Product getProductById(Long id);
+	public Product getProductById(Integer id);
 	
 	public Product addProduct(Product product);
 	
 	public Product updateProduct(Product product);
 	
-	public void deleteProduct(Long id);
+	public void deleteProduct(Integer id);
 
 	Product createProduct(Product product);
 }
