+++ src/main/java/com/coderhouse/repositories/SaleRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.coderhouse.models.Sale;
 
 
-
-public interface SaleRepository extends JpaRepository<Sale, Long>{
+public interface SaleRepository extends JpaRepository<Sale, Integer>{
 
 }
